# docs/swagger/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Monitoring and operations API for Redis Sentinel clusters on Kubernetes.",
        "title": "Redis Sentinel Admin API",
        "contact": {
            "name": "Internal Infra Team"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/config/diff": {
            "get": {
                "description": "Runs CONFIG GET * on every node and returns keys whose values differ across nodes. Only drifted keys are included in the response.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Cross-node config diff",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_operations.ConfigDiff"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "502": {
                        "description": "Node unreachable",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/config/set": {
            "post": {
                "description": "Applies CONFIG SET on a specific node and records the change in the audit log. Requires confirm:true in the request body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "Set a Redis config value",
                "parameters": [
                    {
                        "description": "CONFIG SET parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api_dto.ConfigSetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "confirm not set or invalid request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "502": {
                        "description": "Node unreachable",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/diagnostics/locks": {
            "get": {
                "description": "Scans keys matching a pattern and returns those with no TTL or with a TTL exceeding the stale threshold. Useful for detecting leaked distributed locks.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "diagnostics"
                ],
                "summary": "Stale distributed lock detector",
                "parameters": [
                    {
                        "type": "string",
                        "default": "lock:*",
                        "description": "Key pattern to scan (default lock:*)",
                        "name": "pattern",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 300,
                        "description": "TTL in seconds above which a lock is considered stale; 0 = flag only keys with no TTL (default 300)",
                        "name": "stale_threshold",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_diagnostics.LockReport"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ops/failover": {
            "post": {
                "description": "Performs a lag-check, selects the best replica, then triggers SENTINEL FAILOVER. Use dry_run:true to preview without executing. Requires confirm:true when dry_run is false.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Graceful failover",
                "parameters": [
                    {
                        "description": "Failover options",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api_dto.FailoverRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_operations.FailoverResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "confirm not set",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "No eligible replica or quorum not met",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_duydinhle_redis-sentinel-admin_internal_api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_duydinhle_redis-sentinel-admin_internal_api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_api.APIError"
                }
            }
        },
        "github_com_duydinhle_redis-sentinel-admin_internal_api_dto.ConfigSetRequest": {
            "description": "Redis CONFIG SET parameters",
            "type": "object",
            "properties": {
                "confirm": {
                    "type": "boolean",
                    "example": true
                },
                "key": {
                    "type": "string",
                    "example": "maxmemory-policy"
                },
                "node_addr": {
                    "type": "string",
                    "example": "10.0.0.2:6379"
                },
                "value": {
                    "type": "string",
                    "example": "allkeys-lfu"
                }
            }
        },
        "github_com_duydinhle_redis-sentinel-admin_internal_api_dto.FailoverRequest": {
            "description": "Failover trigger options",
            "type": "object",
            "properties": {
                "confirm": {
                    "type": "boolean",
                    "example": true
                },
                "dry_run": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_duydinhle_redis-sentinel-admin_internal_diagnostics.LockReport": {
            "type": "object",
            "properties": {
                "is_stale": {
                    "description": "true: no TTL or TTL \u003e stale threshold",
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "node_addr": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "ttl_seconds": {
                    "description": "-1 = no TTL set",
                    "type": "integer"
                }
            }
        },
...
        "github_com_duydinhle_redis-sentinel-admin_internal_operations.ConfigDiff": {
            "type": "object",
            "properties": {
                "is_drift": {
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_duydinhle_redis-sentinel-admin_internal_operations.NodeConfigValue"
                    }
                }
            }
        },
        "github_com_duydinhle_redis-sentinel-admin_internal_operations.FailoverResult": {
            "type": "object",
            "properties": {
                "dry_run": {
                    "type": "boolean"
                },
                "elapsed_ms": {
                    "type": "integer"
                },
                "new_master": {
                    "type": "string"
                },
                "old_master": {
                    "type": "string"
                },
                "pre_checks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "selected_replica": {
                    "type": "string"
                }
            }
        },
        "github_com_duydinhle_redis-sentinel-admin_internal_operations.NodeConfigValue": {
            "type": "object",
            "properties": {
                "node_addr": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    }
}